// src/app/shared/model/app-cust-company-fin-data-obj.model.ts
export class AppCustCompanyFinDataObj {
    AppCustCompanyFinDataId: number;
    AppCustId: number;
    GrossMonthlyIncomeAmt: number;
    GrossProfitAmt: number;
    GrossMonthlyExpenseAmt: number;
    ReturnOfInvestmentPrcnt: number;
    ReturnOfEquityPrcnt: number;
    ReturnOfAssetPrcnt: number;
    ProfitMarginPrcnt: number;
    CurrentRatioPrcnt: number;
    DebtEquityRatioPrcnt: number;
    InvTurnOverPrcnt: number;
    ArTurnOverPrcnt: number;
    GrowthPrcnt: number;
    WorkingCapitalAmt: number;
    OthMonthlyInstAmt: number;
    DateAsOf: any;
    Revenue: number;
    OprCost: number;
    ProfitBeforeTax: number;
    CurrAsset: number;
    NetFixedAsset: number;
    TotalAsset: number;
    CurrLiablts: number;
    LongTermLiablts: number;
    ShareholderEquity: number;
    CurrRatio: number;
    RowVersion: string[];

constructor() { 
    this.AppCustCompanyFinDataId = 0;
    this.AppCustId = 0;
    this.GrossMonthlyIncomeAmt = 0;
    this.GrossMonthlyExpenseAmt = 0;
    this.GrossProfitAmt = 0;
    this.ReturnOfInvestmentPrcnt = 0;
    this.ReturnOfEquityPrcnt = 0;
    this.ReturnOfAssetPrcnt = 0;
    this.ProfitMarginPrcnt = 0;
    this.CurrentRatioPrcnt = 0;
    this.DebtEquityRatioPrcnt = 0;
    this.InvTurnOverPrcnt = 0;
    this.ArTurnOverPrcnt = 0;
    this.GrowthPrcnt = 0;
    this.WorkingCapitalAmt = 0;
    this.OthMonthlyInstAmt = 0;
    this.Revenue = 0;
    this.OprCost = 0;
    this.ProfitBeforeTax = 0;
    this.CurrAsset = 0;
    this.NetFixedAsset = 0;
    this.TotalAsset = 0;
    this.CurrLiablts = 0;
    this.LongTermLiablts = 0;
    this.ShareholderEquity = 0;
    this.CurrRatio = 0;
    this.RowVersion = [];
    }
}